added tests

+++ smb-csi/driver/healtCheck.go
@@ -37,13 +37,13 @@ func FsInfo(path string) (int64, int64, int64, int64, int64, int64, error) {
 	}
 
 	// Available is blocks available * fragment size
-	available := int64(statfs.Bavail) * int64(statfs.Bsize)
+	available := int64(statfs.Bavail) * statfs.Bsize
 
 	// Capacity is total block count * fragment size
-	capacity := int64(statfs.Blocks) * int64(statfs.Bsize)
+	capacity := int64(statfs.Blocks) * statfs.Bsize
 
 	// Usage is block being used * fragment size (aka block size).
-	usage := (int64(statfs.Blocks) - int64(statfs.Bfree)) * int64(statfs.Bsize)
+	usage := (int64(statfs.Blocks) - int64(statfs.Bfree)) * statfs.Bsize
 
 	inodes := int64(statfs.Files)
 	inodesFree := int64(statfs.Ffree)
